Refactor WriteFlowController by subclassing it
